refactoring Client/Server classes: both API methods shall be in the same class. Started to migrate that with the "Hello" class. This shall be done until all api.client and api.server Classes are moved to appropriate org.loklak.api sub-packages
