+++ src/org/loklak/LoklakServer.java
@@ -66,6 +66,7 @@ import org.eclipse.jetty.security.ConstraintSecurityHandler;
 import org.eclipse.jetty.security.HashLoginService;
 import org.eclipse.jetty.security.LoginService;
 import org.eclipse.jetty.security.authentication.BasicAuthenticator;
+import org.loklak.api.p2p.Hello;
 import org.loklak.api.server.AccessServlet;
 import org.loklak.api.server.AppsServlet;
 import org.loklak.api.server.AssetServlet;
@@ -73,7 +74,6 @@ import org.loklak.api.server.CampaignServlet;
 import org.loklak.api.server.CrawlerServlet;
 import org.loklak.api.server.DumpDownloadServlet;
 import org.loklak.api.server.GeocodeServlet;
-import org.loklak.api.server.HelloServlet;
 import org.loklak.api.server.PeersServlet;
 import org.loklak.api.server.ProxyServlet;
 import org.loklak.api.server.PushServlet;
@@ -492,7 +492,7 @@ public class LoklakServer {
         servletHandler.addServlet(AccessServlet.class, "/api/access.html");
         servletHandler.addServlet(AccessServlet.class, "/api/access.txt");
         servletHandler.addServlet(AppsServlet.class, "/api/apps.json");
-        servletHandler.addServlet(HelloServlet.class, new HelloServlet().getAPIPath() /*"/api/hello.json"*/);
+        servletHandler.addServlet(Hello.class, new Hello().getAPIPath() /*"/api/hello.json"*/);
         servletHandler.addServlet(PeersServlet.class, "/api/peers.json");
         servletHandler.addServlet(CrawlerServlet.class, "/api/crawler.json");
         servletHandler.addServlet(StatusServlet.class, "/api/status.json");
