--- src/org/loklak/api/server/HelloServlet.java
@@ -1,60 +1,0 @@
-/**
- *  HelloServlet
- *  Copyright 22.02.2015 by Michael Peter Christen, @0rb1t3r
- *
- *  This library is free software; you can redistribute it and/or
- *  modify it under the terms of the GNU Lesser General Public
- *  License as published by the Free Software Foundation; either
- *  version 2.1 of the License, or (at your option) any later version.
- *  
- *  This library is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- *  Lesser General Public License for more details.
- *  
- *  You should have received a copy of the GNU Lesser General Public License
- *  along with this program in the file lgpl21.txt
- *  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package org.loklak.api.server;
-
-import org.json.JSONObject;
-import org.loklak.server.APIHandler;
-import org.loklak.server.APIServiceLevel;
-import org.loklak.server.AbstractAPIHandler;
-import org.loklak.server.Authorization;
-import org.loklak.server.Query;
-
-/**
- * Servlet to span the message peer-to-peer network.
- * This servlet is called to announce the existence of the remote peer.
- */
-public class HelloServlet extends AbstractAPIHandler implements APIHandler {
-    
-    private static final long serialVersionUID = 1839868262296635665L;
-
-
-    @Override
-    public APIServiceLevel getDefaultServiceLevel() {
-        return APIServiceLevel.PUBLIC;
-    }
-
-    @Override
-    public APIServiceLevel getCustomServiceLevel(Authorization auth) {
-        return APIServiceLevel.PUBLIC;
-    }
-
-    @Override
-    public String getAPIPath() {
-        return "/api/hello.json";
-    }
-    
-    @Override
-    public JSONObject serviceImpl(Query call, Authorization rights) {
-        JSONObject json = new JSONObject(true);
-        json.put("status", "ok");
-        return json;
-    }
-    
-}
